Atualizacao de repositorio.

+++ BolaoNet.TestsVS/Business/Facade/BolaoCopaMundo2014FacadeBO.cs
@@ -125,55 +125,55 @@ namespace BolaoNet.TestsVS.Business.Facade
             #region Boloes Criterios Pontos
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)1) { Pontos = 2 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)1) { Pontos = 2 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)2) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)2) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)3) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)3) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)4) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)4) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)5) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)5) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)6) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)6) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)7) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)7) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)8) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)8) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)9) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)9) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)10) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)10) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)11) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)11) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)12) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)12) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)13) { Pontos = 0 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)13) { Pontos = 0 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)14) { Pontos = -2 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)14) { Pontos = -2 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)15) { Pontos = 1 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)15) { Pontos = 1 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)16) { Pontos = 1 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)16) { Pontos = 1 });
 
             StoreData<Entities.Boloes.BolaoCriterioPontos>(_bolaoCriterioPontosBO,
-                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (Entities.Boloes.BolaoCriterioPontos.CriteriosID)17) { Pontos = 5 });
+                new Entities.Boloes.BolaoCriterioPontos(bolao.Nome, (int)(Entities.Boloes.BolaoCriterioPontos.CriteriosID)17) { Pontos = 5 });
 
             #endregion
